QTextDocument: avoid relocations
Hand-roll a string table here, since there's opportunity for
sharing string data between string entries.

Effects on Linux AMD64 GCC 4.9-trunk release stripped:
  text:   -88B
  data:   -64B
  relocs:  -5

Maintainability is hurt somewhat, but it is expected that the
contents of this string table do not change much, so the
overall effect is still very low.

Change-Id: I2a22da4c8548c53ef31c33319b4652f3cb6f62f5
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>