Restore handling of bullets for lists with a font size of > 36 pixels
This patch partially reverts ad443dfb1d8e9096c4913686aa2ed0bc9b3f5de7,
the test that was added is still there as it can be used to compare
against Chrome and to check that indenting works correctly.

The original behavior of clipping the bullets is correct because the
specified indent is where the text itself should be positioned, the bullet
uses the available space to the side of it. This is how other web browsers
handle the same situation.

Change-Id: I63440e037d8efec356459c09228ef4817ccb9cb6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>