don't emit comments to generated qfeatures.h
the file is not meant to be human-readable and even less editable, so
there is no point in having comments in it.

also, it was completely inconsistent to start with: features without
dependencies were listed as "templates" in form of commented out
 #defines to disable them manually, while features with dependencies
had a respective #ifdef block to be automatically disabled, but no
"template" to disable them manually. now only the #ifdef blocks remain.
the actual configuration is done by configure via qconfig.h.

Change-Id: I8b9e56ba570908dad4cc6dfcd24bf0e1da8b290f
Reviewed-by: Tasuku Suzuki <stasuku@gmail.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>