Fail when QT_POINTER_SIZE is not set
Some cross compile setups don't have special treatment
and don't have a configure check. To notice this early
in the build fail at this state. One example is to cross
compile from Windows for eg QNX or even linux embedded.

If QT_POINTER_SIZE is not set, later v4 build in
qtdeclarative is failing with a unmeaningful compile
error.

Change-Id: Iff9498d916371241b0b044b9543f9cd2946ecf42
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>